Settings: Unlock non-secure users automatically
...while moving apps or media content from internal storage to
adopted storage and vice-versa.
Currently, for FBE enabled devices, whenever move is triggered for
either app or media content, there is a prompt for user to enter
their credentails in order to unlock storage if the user has one.
But if the user does not have a authentication set, the wizard page
is stuck. This leads to a bad user experience.

StorageWizardMigrateConfirm & StorageWizardMoveConfirm uses
ChooseLockSettingsHelper to launch the authentication method set
by the user in order to unlock the storage. But, it does not handle
the use case where user has no authentication set, causing no
return to the wizard activities, thus blocking the flow. In order
to solve that, checking if the user is not secure, attempt unlock
with null credentials to unlock the storage otherwise prompt the
user to unlock.

Test: manual
Steps:
1. Create multiple users with atleast one user with an authentication
   method setup and one without authentication (apart from system user)
2. Install same apps for all users and also add some content to the
   media storage.
3. Format an SD Card as adopted.
4. Attempt to move content.
5. Check that the wizard is not stuck and it is prompting the locked
   users to authenticate.
6. Reboot
7. After reboot, Go to Settings > All Apps > app_name
   > Storage & Cache
8. Tap on "Change" under Storage and Select Adopted storage.
9. Check the same as Step#5

Change-Id: Ib72e9073386457711cc1d2ba745a24cbea91cd8a
Signed-off-by: Dmitrii <bankersenator@gmail.com>